+++ include/MCUC.hpp
@@ -7,7 +7,7 @@
 #include <core/io/GPIO.hpp>
 #include <core/io/pin.hpp>
 #include <core/io/types/CANMessage.hpp>
-#include <dev/PowertrainCAN.hpp>
+#include <PowertrainCAN.hpp>
 #include <models/MCuC_Model.hpp>
 
 namespace io = core::io;
